- Refactored file locations in Services project
