+++ FrannHammer.Services.Tests/BaseMoveHitboxSearchPredicateServiceTest.cs
@@ -1,4 +1,5 @@
 ﻿using FrannHammer.Models;
+using FrannHammer.Services.MoveSearch;
 using NUnit.Framework;
 
 namespace FrannHammer.Services.Tests
