+++ FrannHammer.Services.Tests/Harnesses/MetadataHarness.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using FrannHammer.Models;
+using FrannHammer.Services.MoveSearch;
 using FrannHammer.Services.Tests.Suts;
 using NUnit.Framework;
