+++ FrannHammer.Services.Tests/Harnesses/MoveDataHarness.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq;
 using FrannHammer.Models;
+using FrannHammer.Services.MoveSearch;
 using FrannHammer.Services.Tests.Suts;
 using NUnit.Framework;
