+++ FrannHammer.Services/MoveSearch/RangeMatchProcessingService.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using FrannHammer.Models;
 
-namespace FrannHammer.Services
+namespace FrannHammer.Services.MoveSearch
 {
     public class RangeMatchProcessingService
     {
