+++ FrannHammer.Services/MoveSearch/SearchPredicateService.cs
@@ -2,7 +2,7 @@
 using FrannHammer.Core;
 using FrannHammer.Models;
 
-namespace FrannHammer.Services
+namespace FrannHammer.Services.MoveSearch
 {
     public class SearchPredicateService
     {
